+++ src/main/java/com/github/afanas10101111/jtb/bot/JavaRushTelegramBot.java
@@ -1,6 +1,8 @@
 package com.github.afanas10101111.jtb.bot;
 
+import com.github.afanas10101111.jtb.client.GroupClient;
 import com.github.afanas10101111.jtb.command.CommandContainer;
+import com.github.afanas10101111.jtb.service.GroupSubService;
 import com.github.afanas10101111.jtb.service.SendBotMessageServiceImpl;
 import com.github.afanas10101111.jtb.service.UserService;
 import lombok.Setter;
@@ -9,6 +11,9 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.meta.api.objects.Update;
+
+import static com.github.afanas10101111.jtb.bot.util.BotUpdateUtil.extractCommandIdentifier;
+import static com.github.afanas10101111.jtb.bot.util.BotUpdateUtil.isMessageExists;
 
 @Slf4j
 @Setter
@@ -20,15 +25,15 @@ public class JavaRushTelegramBot extends TelegramLongPollingBot {
 
     private final CommandContainer commandContainer;
 
-    public JavaRushTelegramBot(UserService userService) {
-        commandContainer = new CommandContainer(new SendBotMessageServiceImpl(this), userService);
+    public JavaRushTelegramBot(UserService userService, GroupSubService groupSubService, GroupClient client) {
+        commandContainer = new CommandContainer(new SendBotMessageServiceImpl(this), userService, groupSubService, client);
     }
 
     @Override
     public void onUpdateReceived(Update update) {
-        if(update.hasMessage() && update.getMessage().hasText()) {
+        if(isMessageExists(update)) {
             commandContainer
-                    .retrieveCommand(update.getMessage().getText().trim().split(" ")[0].toLowerCase())
+                    .retrieveCommand(extractCommandIdentifier(update))
                     .execute(update);
         }
     }