+++ src/main/java/com/github/afanas10101111/jtb/client/dto/GroupInfoType.java
@@ -1,0 +1,11 @@
+package com.github.afanas10101111.jtb.client.dto;
+
+public enum GroupInfoType {
+    UNKNOWN,
+    CITY,
+    COMPANY,
+    COLLEGE,
+    TECH,
+    SPECIAL,
+    COUNTRY
+}