+++ src/main/java/com/github/afanas10101111/jtb/client/dto/MeGroupInfo.java
@@ -1,0 +1,9 @@
+package com.github.afanas10101111.jtb.client.dto;
+
+import lombok.Data;
+
+@Data
+public class MeGroupInfo {
+    private MeGroupInfoStatus status;
+    private Integer userGroupId;
+}