+++ src/main/java/com/github/afanas10101111/jtb/command/StartCommand.java
@@ -6,6 +6,8 @@ import com.github.afanas10101111.jtb.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
+import static com.github.afanas10101111.jtb.bot.util.BotUpdateUtil.extractChatId;
+
 @RequiredArgsConstructor
 public class StartCommand implements Command {
     public static final String MESSAGE = "Привет! Я готов помочь тебе быть в курсе выхода всех интересующих тебя статей на Java Rush";
@@ -15,7 +17,7 @@ public class StartCommand implements Command {
 
     @Override
     public void execute(Update update) {
-        String chatId = update.getMessage().getChatId().toString();
+        String chatId = extractChatId(update);
         userService.findByChatId(chatId).ifPresentOrElse(
                 u -> {
                     u.setActive(true);
@@ -28,6 +30,6 @@ public class StartCommand implements Command {
                     userService.save(user);
                 }
         );
-        messageService.sendMessage(update.getMessage().getChatId().toString(), MESSAGE);
+        messageService.sendMessage(chatId, MESSAGE);
     }
 }