+++ src/main/java/com/github/afanas10101111/jtb/command/StopCommand.java
@@ -5,6 +5,8 @@ import com.github.afanas10101111.jtb.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
+import static com.github.afanas10101111.jtb.bot.util.BotUpdateUtil.extractChatId;
+
 @RequiredArgsConstructor
 public class StopCommand implements Command {
     public static final String MESSAGE = "До свидания! Все твои подписки деактивированы";
@@ -14,8 +16,9 @@ public class StopCommand implements Command {
 
     @Override
     public void execute(Update update) {
-        messageService.sendMessage(update.getMessage().getChatId().toString(), MESSAGE);
-        userService.findByChatId(update.getMessage().getChatId().toString()).ifPresent(
+        String chatId = extractChatId(update);
+        messageService.sendMessage(chatId, MESSAGE);
+        userService.findByChatId(chatId).ifPresent(
                 u -> {
                     u.setActive(false);
                     userService.save(u);