+++ src/main/java/com/github/afanas10101111/jtb/command/UnknownCommand.java
@@ -4,6 +4,8 @@ import com.github.afanas10101111.jtb.service.SendBotMessageService;
 import lombok.RequiredArgsConstructor;
 import org.telegram.telegrambots.meta.api.objects.Update;
 
+import static com.github.afanas10101111.jtb.bot.util.BotUpdateUtil.extractChatId;
+
 @RequiredArgsConstructor
 public class UnknownCommand implements Command {
     public static final String MESSAGE = "Ничего не понял :(\nВведи /help чтобы ознакомиться со списком команд";
@@ -12,6 +14,6 @@ public class UnknownCommand implements Command {
 
     @Override
     public void execute(Update update) {
-        service.sendMessage(update.getMessage().getChatId().toString(), MESSAGE);
+        service.sendMessage(extractChatId(update), MESSAGE);
     }
 }