+++ src/main/java/com/github/afanas10101111/jtb/model/GroupSub.java
@@ -1,0 +1,43 @@
+package com.github.afanas10101111.jtb.model;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToMany;
+import javax.persistence.Table;
+import java.util.HashSet;
+import java.util.Set;
+
+@Data
+@EqualsAndHashCode(of = "id")
+@Entity
+@Table(name = "group_sub")
+public class GroupSub {
+
+    @Id
+    private Integer id;
+
+    @Column(name = "title")
+    private String title;
+
+    @Column(name = "last_article_id")
+    private Integer lastArticleId;
+
+    @ManyToMany(fetch = FetchType.EAGER)
+    @JoinTable(
+            name = "group_to_user",
+            joinColumns = @JoinColumn(name = "group_id"),
+            inverseJoinColumns = @JoinColumn(name = "user_id")
+    )
+    private Set<User> users = new HashSet<>();
+
+    public void addUser(User user) {
+        users.add(user);
+    }
+}