+++ src/main/java/com/github/afanas10101111/jtb/model/User.java
@@ -1,13 +1,19 @@
 package com.github.afanas10101111.jtb.model;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.Id;
+import javax.persistence.ManyToMany;
+import javax.persistence.OrderBy;
 import javax.persistence.Table;
+import java.util.Set;
 
 @Data
+@EqualsAndHashCode(of = "chatId")
 @Entity
 @Table(name = "tg_user")
 public class User {
@@ -18,4 +24,8 @@ public class User {
 
     @Column(name = "active")
     private boolean active;
+
+    @ManyToMany(mappedBy = "users", fetch = FetchType.EAGER)
+    @OrderBy("id")
+    private Set<GroupSub> groupSubs;
 }