+++ src/main/java/com/github/afanas10101111/jtb/service/GroupSubService.java
@@ -1,0 +1,8 @@
+package com.github.afanas10101111.jtb.service;
+
+import com.github.afanas10101111.jtb.client.dto.GroupDiscussionInfo;
+import com.github.afanas10101111.jtb.model.GroupSub;
+
+public interface GroupSubService {
+    GroupSub save(String chatId, GroupDiscussionInfo groupDiscussionInfo);
+}