+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## 0.5.0-SNAPSHOT
+* [FEATURE] JTB-4: Add logic for adding the group to the subscription
+* [FEATURE] JTB-5: Add command to see list of my group subscriptions
+
 ## 0.4.0-SNAPSHOT
 * [FEATURE] JTB-0: Add repository layer